Fix: Wolf Unit synchronization for midway hitbox and input accura
Vblank timing correction was applied to wolfUnitFrame(). 
The main issue was the imprecise vblank calculation in the shared driver for midway games. 
Lack of a proper scanline offset caused the game's input reading and collision detection.
The correction applied is in the latest fbneo which was exported to fightcade fbneo
  "INT32 our_line = (i + 274) % 289;"
the change restores accurate input timing and
resolves known hitbox and input drop issues in midway games, significantly restores accuracy to 
the entire Wolf Unit games.